requirements.txt
pandas==1.3.3
streamlit==0.88.0
numpy==1.19.4
regex==2021.7.6
pymongo==3.11.4
python_freeDictionaryAPI==0.9.10
